+++ ready/AzNamingTool/Models/SiteConfiguration.cs
@@ -1,11 +1,11 @@
-﻿namespace AzureNamingTool.Helpers
+﻿namespace AzureNamingTool.Models
 {
-    public class Config
+    public class SiteConfiguration
     {
         public string? SALTKey { get; set; }
         public string? AdminPassword { get; set; }
         public string? APIKey { get; set; }
         public string? AppTheme { get; set; }
-        public bool? DevMode{ get; set; } = false;
+        public bool? DevMode { get; set; } = false;
     }
 }