+++ ready/AzNamingTool/Services/CustomComponentService.cs
@@ -12,7 +12,7 @@ namespace AzureNamingTool.Services
             try
             {
                 // Get list of items
-                var items = await GeneralHelper.GetList<CustomComponent>();
+                var items = await ConfigurationHelper.GetList<CustomComponent>();
                 serviceResponse.ResponseObject = items.OrderBy(x => x.SortOrder).ToList();
                 serviceResponse.Success = true;
             }
@@ -30,7 +30,7 @@ namespace AzureNamingTool.Services
             try
             {
                 // Get list of items
-                var items = await GeneralHelper.GetList<CustomComponent>();
+                var items = await ConfigurationHelper.GetList<CustomComponent>();
                 serviceResponse.ResponseObject = items.Where(x => x.ParentComponent == parenttype).OrderBy(x => x.SortOrder).ToList();
                 serviceResponse.Success = true;
             }
@@ -48,7 +48,7 @@ namespace AzureNamingTool.Services
             try
             {
                 // Get list of items
-                var data = await GeneralHelper.GetList<CustomComponent>();
+                var data = await ConfigurationHelper.GetList<CustomComponent>();
                 var item = data.Find(x => x.Id == id);
                 serviceResponse.ResponseObject = item;
                 serviceResponse.Success = true;
@@ -67,7 +67,7 @@ namespace AzureNamingTool.Services
             try
             {
                 // Make sure the new item short name only contains letters/numbers
-                if (!GeneralHelper.CheckAlphanumeric(item.ShortName))
+                if (!ValidationHelper.CheckAlphanumeric(item.ShortName))
                 {
                     serviceResponse.Success = false;
                     serviceResponse.ResponseObject = "Short name must be alphanumeric.";
@@ -78,7 +78,7 @@ namespace AzureNamingTool.Services
                 item.ShortName = item.ShortName.ToLower();
 
                 // Get list of items
-                var items = await GeneralHelper.GetList<CustomComponent>();
+                var items = await ConfigurationHelper.GetList<CustomComponent>();
 
                 // Set the new id
                 if (item.Id == 0)
@@ -144,7 +144,7 @@ namespace AzureNamingTool.Services
                 }
 
                 // Write items to file
-                await GeneralHelper.WriteList<CustomComponent>(items);
+                await ConfigurationHelper.WriteList<CustomComponent>(items);
                 serviceResponse.ResponseObject = "Item added!";
                 serviceResponse.Success = true;
             }
@@ -162,7 +162,7 @@ namespace AzureNamingTool.Services
             try
             {
                 // Get list of items
-                var items = await GeneralHelper.GetList<CustomComponent>();
+                var items = await ConfigurationHelper.GetList<CustomComponent>();
                 // Get the specified item
                 var item = items.Find(x => x.Id == id);
                 // Remove the item from the collection
@@ -177,7 +177,7 @@ namespace AzureNamingTool.Services
                 }
 
                 // Write items to file
-                await GeneralHelper.WriteList<CustomComponent>(items);
+                await ConfigurationHelper.WriteList<CustomComponent>(items);
                 serviceResponse.Success = true;
             }
             catch (Exception ex)
@@ -201,7 +201,7 @@ namespace AzureNamingTool.Services
                 foreach (CustomComponent item in items)
                 {
                     // Make sure the new item short name only contains letters/numbers
-                    if (!GeneralHelper.CheckAlphanumeric(item.ShortName))
+                    if (!ValidationHelper.CheckAlphanumeric(item.ShortName))
                     {
                         serviceResponse.Success = false;
                         serviceResponse.ResponseObject = "Short name must be alphanumeric.";
@@ -218,7 +218,7 @@ namespace AzureNamingTool.Services
                 }
 
                 // Write items to file
-                await GeneralHelper.WriteList<CustomComponent>(newitems);
+                await ConfigurationHelper.WriteList<CustomComponent>(newitems);
                 serviceResponse.Success = true;
             }
             catch (Exception ex)